refactor: extension 파일 통합, NameSpace rename 리팩토링
