+++ Expo1900/Expo1900/Controller/ExhibitionWorkListViewController.swift
@@ -26,7 +26,7 @@ final class ExhibitionWorkListViewController: UIViewController {
     func setUpJSONData() {
         let jsonDecoder: JSONDecoder = JSONDecoder()
         
-        guard let dataAsset: NSDataAsset = NSDataAsset(name: NameSpace.itemKeys) else { return }
+        guard let dataAsset: NSDataAsset = NSDataAsset(name: ExpositionNameSpace.itemKeys) else { return }
         
         do {
             exhibitionWork = try jsonDecoder.decode([ExhibitionWork].self, from: dataAsset.data)
@@ -61,5 +61,26 @@ extension ExhibitionWorkListViewController: UITableViewDelegate {
         detailWorkVC.exhibitionWork = self.exhibitionWork[indexPath.row]
         
         self.navigationController?.pushViewController(detailWorkVC, animated: true)
+    }
+}
+
+extension ExhibitionWorkListViewController {
+    func setUpTableView() {
+        self.exhibitionWorkTableView.register(TableViewCell.self, forCellReuseIdentifier: TableViewCell.cellIdentifier)
+        self.view.addSubview(self.exhibitionWorkTableView)
+        
+        self.exhibitionWorkTableView.translatesAutoresizingMaskIntoConstraints = false
+        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
+                                                   attribute: .top, relatedBy: .equal, toItem: self.view, attribute: .top,
+                                                   multiplier: 1.0, constant: 0))
+        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
+                                                   attribute: .bottom, relatedBy: .equal, toItem: self.view,
+                                                   attribute: .bottom, multiplier: 1.0, constant: 0))
+        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
+                                                   attribute: .leading, relatedBy: .equal, toItem: self.view,
+                                                   attribute: .leading, multiplier: 1.0, constant: 0))
+        self.view.addConstraint(NSLayoutConstraint(item: self.exhibitionWorkTableView,
+                                                   attribute: .trailing, relatedBy: .equal, toItem: self.view,
+                                                   attribute: .trailing, multiplier: 1.0, constant: 0))
     }
 }
